requirements.txt
-e git+https://github.com/django/django.git@1.7.7#egg=django
MySQL-python==1.2.5
virtualenv==12.0.5
django-tastypie==0.12.1